Simplified project structure
As more actions are added it's become apparent that registering all
actions in the "src/actions/index.js" file and using them through that
file is going to be cleaner than requiring them individually from the
other action files directly. It also has the added benefit of ensuring
that the index file is updated as new actions are added.