Add slight delay to news refresh action
This is being added solely to work around a bug in React Native's
refresh control. When the SpaceWatch app is in debug mode, the fetch
text action returns synchronously, and apparently the refresh control
relies on a delay between being toggled on and off, because it gets
stuck in the "on" position in the synchronous version of events.