Add built in type constructors to signal default values.
  As the Model class was being used it was noticed that empty default
values were being used a lot. For instance, an empty string literal ("")
was being used to signify the String type a lot. Instead of doing that,
Model now takes the String constructor itself to signify just using an
empty string. This is either better or worse, it looks better, but is
implying some sort of type validation, which it isn't doing yet. So I've
added a note to do that later and am moving on for now.