Fix issue with Notifier syncing properly with store.
  The React Native notification interface is very simple and does not
have an "exists" method. This causes an issue where we cannot know if
the hidden native state already has a notification registered or not,
which makes actions like "toggle" impossible. To combat this the
Notifier interface was simplified and syncing responsibilities were
pushed back into the reducers as well as higher up the stack in the
front end applications themselves. This isn't ideal, but it's faster
than stopping to hack on React Native right now.