Change project name.
This is just to ensure that the project gets a cheap URL. Maison, if you
were wondering, is 'house' in French, which makes very little sense for
a news aggregator about space, but the domain name is 7 bucks a year,
and the American English pronunciation of space.maison sounds nice if
you pronounce it like space.mason, which also wouldn't make any sense,
but still sounds nice. This commit message reads like a bad parody of
Nate DiMeo's Memory Palace.